Fixed x64 compilation
git-svn-id: http://miranda.googlecode.com/svn/trunk@10197 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/IRCG/options.cpp
@@ -80,7 +80,7 @@ void CIrcProto::ReadSettings( TDbSetting* sets, int count )
 						if ( p->defStr == NULL )
 							*ptr = 0;
 						else 
-							lstrcpyn(( TCHAR* )ptr, p->defStr, p->size );
+							lstrcpyn(( TCHAR* )ptr, p->defStr, (int)p->size );
 					}
 					else *( TCHAR** )ptr = mir_tstrdup( p->defStr );
 				}
